number tokenzier fix for alphanumeric values

+++ COLID.SearchService.Services/Implementation/IndexService.cs
@@ -10,6 +10,7 @@ using COLID.SearchService.DataModel.Search;
 using COLID.SearchService.Repositories.Interface;
 using COLID.SearchService.Services.Interface;
 using Elasticsearch.Net;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Newtonsoft.Json;
@@ -25,12 +26,17 @@ namespace COLID.SearchService.Services.Implementation
         private readonly IElasticSearchRepository _elasticSearchRepository;
         private readonly ColidMessageQueueOptions _mqOptions;
         private readonly ILogger<IndexService> _logger;
+        private readonly bool _reindexingSwitch;
+        private readonly IConfiguration _configuration;
 
-        public IndexService(IElasticSearchRepository repository, IOptionsMonitor<ColidMessageQueueOptions> messageQueuingOptionsAccessor, ILogger<IndexService> logger)
+        public IndexService(IElasticSearchRepository repository, IOptionsMonitor<ColidMessageQueueOptions> messageQueuingOptionsAccessor, ILogger<IndexService> logger, IConfiguration configuration)
         {
             _elasticSearchRepository = repository;
             _mqOptions = messageQueuingOptionsAccessor.CurrentValue;
             _logger = logger;
+            _configuration = configuration;
+            _reindexingSwitch = _configuration.GetValue<bool>("ReindexSwitch");
+            _logger.LogInformation($"ReindexSwitch is allowed {_reindexingSwitch}", _reindexingSwitch);
         }
 
         public IDictionary<string, Action<string>> OnTopicReceivers => new Dictionary<string, Action<string>>() {
@@ -39,43 +45,46 @@ namespace COLID.SearchService.Services.Implementation
 
         public async void ReindexingSwitch(string pidUriString)
         {
-            //var document = (JObject)JsonConvert.DeserializeObject(pidUriString);
-            //var lastPidUris = document["lastPidUris"];
-            //bool continueCheck = true;
-            //DateTime loopStart = DateTime.Now;
-            //_logger.LogInformation($"Checking if the piduris have been received in new index at {loopStart} hours", loopStart);
-            //int myCount = 0;
-            //while (continueCheck && DateTime.Now.Subtract(loopStart).Hours < 4)
-            //{
-            //    myCount++;
-            //    _logger.LogInformation($"The loop is running for {myCount} time", myCount);
-            //    lastPidUris.ToList().ForEach(pidUri =>
-            //    {
-            //        try
-            //        {
-            //            var response = _elasticSearchRepository.GetDocument(HttpUtility.UrlEncode(pidUri.ToString()), UpdateIndex.Published);
-            //            if (response != null && continueCheck)
-            //            {
-            //                _logger.LogInformation($"Document present {pidUri} in new index", pidUri);
+            if (_reindexingSwitch)
+            {
+            _logger.LogInformation($"Reindexing switch is true for non local env and thus we wait before switching index");
+            var document = (JObject)JsonConvert.DeserializeObject(pidUriString);
+            var lastPidUris = document["lastPidUris"];
+            bool continueCheck = true;
+            DateTime loopStart = DateTime.Now;
+            _logger.LogInformation($"Checking if the piduris have been received in new index at {loopStart} hours", loopStart);
+            int myCount = 0;
+            while (continueCheck && DateTime.Now.Subtract(loopStart).Hours < 4)
+            {
+                myCount++;
+                _logger.LogInformation($"The loop is running for {myCount} time", myCount);
+                lastPidUris.ToList().ForEach(pidUri =>
+                {
+                    try
+                    {
+                        var response = _elasticSearchRepository.GetDocument(HttpUtility.UrlEncode(pidUri.ToString()), UpdateIndex.Published);
+                        if (response != null && continueCheck)
+                        {
+                            _logger.LogInformation($"Document present {pidUri} in new index", pidUri);
 
-            //                continueCheck = false;
-            //            }
-            //        }
-            //        catch (System.Exception ex)
-            //        {
-            //            if (ex is EntityNotFoundException)
-            //            {
-            //                _logger.LogWarning(ex, $"Document not recieved yet in new index for {pidUri} by message queue", pidUri.ToString());
-            //            }
-            //        }
-            //    });
+                            continueCheck = false;
+                        }
+                    }
+                    catch (System.Exception ex)
+                    {
+                        if (ex is EntityNotFoundException)
+                        {
+                            _logger.LogWarning(ex, $"Document not recieved yet in new index for {pidUri} by message queue", pidUri.ToString());
+                        }
+                    }
+                });
 
-            //    await Task.Delay(600000);
+                await Task.Delay(600000);
 
-            //}
-            _logger.LogInformation("This method is commented for Open Source Version");
-            //_logger.LogInformation($"Loop finished. Switching Search Aliases at {DateTime.Now} hours");
-            //SwitchAndDeleteOldIndex();
+            }
+            _logger.LogInformation($"Loop finished. Switching Search Aliases at {DateTime.Now} hours");
+            SwitchAndDeleteOldIndex();
+            }
         }
 
         private void SwitchAndDeleteOldIndex()
@@ -147,16 +156,22 @@ namespace COLID.SearchService.Services.Implementation
                     _elasticSearchRepository.IndexMetadata(JObject.FromObject(metadataObject));
                 }
                 _elasticSearchRepository.UpdateMetadataSearchAlias(rollbackActions);
-                _elasticSearchRepository.UpdateDocumentSearchAlias(rollbackActions, UpdateIndex.Draft, SearchIndex.Draft);
-                _elasticSearchRepository.UpdateDocumentSearchAlias(rollbackActions, UpdateIndex.Published, SearchIndex.Published);
+                //_elasticSearchRepository.UpdateDocumentSearchAlias(rollbackActions, UpdateIndex.Draft, SearchIndex.Draft);
+                //_elasticSearchRepository.UpdateDocumentSearchAlias(rollbackActions, UpdateIndex.Published, SearchIndex.Published);
 
                 /// At this point the actual indexing process is completed.
                 /// The following processes are there to clean up the elastic.
                 /// It is ignored whether the deletion was successful or not.
-                var oldIndices = oldDraftDocumentIndices.Concat(oldPublishedDocumentIndices).Concat(oldMetadataIndices);
-                foreach (var oldIndex in oldIndices)
+                if (!_reindexingSwitch)
                 {
-                    _elasticSearchRepository.DeleteIndex(oldIndex);
+                    _logger.LogInformation($"We do not need to wait for switching index in Local environment");
+                    _elasticSearchRepository.UpdateDocumentSearchAlias(rollbackActions, UpdateIndex.Draft, SearchIndex.Draft);
+                    _elasticSearchRepository.UpdateDocumentSearchAlias(rollbackActions, UpdateIndex.Published, SearchIndex.Published);
+                    var oldIndices = oldDraftDocumentIndices.Concat(oldPublishedDocumentIndices).Concat(oldMetadataIndices);
+                    foreach (var oldIndex in oldIndices)
+                    {
+                        _elasticSearchRepository.DeleteIndex(oldIndex);
+                    }
                 }
             }
             catch (System.Exception ex)
